Add PID to S3 log name
Summary:
Without this change, when two instances of a single executable run on
the same machine, they use the same logfile.  When one exits, it deletes the
logfile, which causes panics in the other instance.  Incorporating the PID
avoids that.

Test Plan:
Unit tests; also made the same change manually in Spade and
confirmed it works both in standard mode and in replay mode (which by default
starts two instances per machine).

Reviewers: jackgao

Reviewed By: jackgao

Subscribers: #scieng

Differential Revision: https://phabricator.internal.justin.tv/D4331